+++ base/gxband.h
@@ -52,8 +52,8 @@ typedef struct gx_band_page_info_s {
     char bfname[gp_file_name_sizeof];	/* block file name */
     clist_file_ptr bfile;	/* block file, normally 0 */
     const clist_io_procs_t *io_procs;
-    uint tile_cache_size;	/* size of tile cache */
-    ulong line_ptrs_offset;      /* Offset of line_ptrs within tile cache */
+    size_t tile_cache_size;	/* size of tile cache */
+    size_t line_ptrs_offset;      /* Offset of line_ptrs within tile cache */
     int64_t bfile_end_pos;		/* ftell at end of bfile */
     gx_band_params_t band_params;  /* parameters used when writing band list */
                                 /* (actual values, no 0s) */
